Move `origins_match()` to `webauthn-rs-proto`.
`webauthn-rs-core` depends on cryptography libraries, which aren't required
for some configurations of `webauthn-authenticator-rs` (eg: on Windows).
