+++ webauthn-authenticator-rs/Cargo.toml
@@ -72,7 +72,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 [dependencies]
 base64.workspace = true
 base64urlsafedata.workspace = true
-webauthn-rs-proto.workspace = true
+webauthn-rs-proto = { workspace = true, features = ["origin"] }
 webauthn-rs-core = { workspace = true, optional = true }
 
 tracing.workspace = true
