+++ webauthn-authenticator-rs/src/lib.rs
@@ -103,16 +103,14 @@ extern crate num_derive;
 #[macro_use]
 extern crate tracing;
 
-use std::str::FromStr;
-
 use crate::error::WebauthnCError;
 use base64::engine::general_purpose::URL_SAFE_NO_PAD as BASE64_ENGINE;
 use url::Url;
 
-use webauthn_rs_core::WebauthnCore;
 use webauthn_rs_proto::{
-    CreationChallengeResponse, PublicKeyCredential, PublicKeyCredentialCreationOptions,
-    PublicKeyCredentialRequestOptions, RegisterPublicKeyCredential, RequestChallengeResponse,
+    origin::origins_match, CreationChallengeResponse, PublicKeyCredential,
+    PublicKeyCredentialCreationOptions, PublicKeyCredentialRequestOptions,
+    RegisterPublicKeyCredential, RequestChallengeResponse,
 };
 
 pub mod prelude {
@@ -292,7 +290,7 @@ where
                 WebauthnCError::Security
             })?;
 
-        if !WebauthnCore::origins_match(true, true, &origin, &rp_id_url) {
+        if !origins_match(true, true, &origin, &rp_id_url) {
             error!(
                 "Relying party ID ({rp_id_url}) is not a suffix of the effective domain ({origin})"
             );
@@ -367,7 +365,7 @@ where
                 WebauthnCError::Security
             })?;
 
-        if !WebauthnCore::origins_match(true, true, &origin, &rp_id_url) {
+        if !origins_match(true, true, &origin, &rp_id_url) {
             error!(
                 "Relying party ID ({rp_id_url}) is not a suffix of the effective domain ({origin})"
             );
