+++ webauthn-rs-proto/src/lib.rs
@@ -12,6 +12,9 @@ pub mod cose;
 pub mod extensions;
 pub mod options;
 
+#[cfg(any(test, feature = "origin"))]
+pub mod origin;
+
 #[cfg(feature = "wasm")]
 pub mod wasm;
